--- frontend/src/main/java/com/example/utils/CheckedFunction.java
@@ -1,6 +1,0 @@
-package com.example.utils;
-
-@FunctionalInterface
-public interface CheckedFunction {
-    void apply() throws NoNetworkException;
-}
