+++ frontend/src/main/java/com/example/utils/NetworkThrowingFunction.java
@@ -1,0 +1,8 @@
+package com.example.utils;
+
+import com.example.utils.exeptions.NoNetworkException;
+
+@FunctionalInterface
+public interface NetworkThrowingFunction {
+    void apply() throws NoNetworkException;
+}
