+++ frontend/src/main/java/com/example/utils/exeptions/IncorrectCodeExeption.java
@@ -1,4 +1,4 @@
-package com.example.utils;
+package com.example.utils.exeptions;
 
 public class IncorrectCodeExeption extends Exception {
     public IncorrectCodeExeption(String message){
