+++ frontend/src/main/java/com/example/utils/exeptions/NoNetworkException.java
@@ -1,4 +1,4 @@
-package com.example.utils;
+package com.example.utils.exeptions;
 
 public class NoNetworkException extends Exception {
     public NoNetworkException(String message, Throwable cause) {
